# src/main/resources/swagger/poc-sgm-geo.yaml
swagger: "2.0"
info:
  description: "Georeferenced Municipal Information"
  version: "1.0.0"
  title: poc-sgm-geo
  contact:
    email: "lucas.pradella@hotmail.com"
host: "localhost:8090"
basePath: "/v1"
tags:
  - name: "Georeferenced"
    description: "general purpose endpoint"
  - name: "Georeferenced for administrators"
    description: "Endpoint for restricted use by administrators and employees"
  - name: "Georeferenced for applications"
    description: "Endpoints for exclusive use of other applications"
schemes:
  - "https"
  - "http"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header

paths:
  /city-data/{idLocation}/info:
    get:
      security:
        - Bearer: []
      tags:
        - "Georeferenced"
      summary: "Returns the main news"
      description: "USER ROLE is required to access"
      operationId: "searchData"
      consumes:
        - "application/json"
      parameters:
        - name: "idLocation"
          in: "path"
          description: "ID for IBGE"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/InfoResponse"
        "400":
          description: "bady request"
        "404":
          description: "not found"
        "500":
          description: "server error"
          schema:
            $ref: "#/definitions/ErrorResponse"

  /city-data/{idLocation}/metadata:
    get:
      security:
        - Bearer: []
      tags:
        - "Georeferenced for administrators"
      summary: "Returns the main news"
      description: "ROLE USER and ADMIN is required to access"
      operationId: "metadata"
      consumes:
        - "application/json"
      parameters:
        - name: "idLocation"
          in: "path"
          description: "ID for IBGE"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/MetadataResponse"
        "400":
          description: "bady request"
        "404":
          description: "not found"
        "500":
          description: "server error"
          schema:
            $ref: "#/definitions/ErrorResponse"

  /geographic-coordinates/{idLocation}:
    get:
      security:
        - Bearer: []
      tags:
        - "Georeferenced for applications"
      summary: "Returns the main news"
      description: "ROLE APPLICATION and NOT USER is required to access"
      operationId: "coordinates"
      consumes:
        - "application/json"
      parameters:
        - name: "idLocation"
          in: "path"
          description: "ID for IBGE"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/CoordinatesResponse"
        "400":
          description: "bady request"
        "404":
          description: "not found"
        "500":
          description: "server error"
          schema:
            $ref: "#/definitions/ErrorResponse"

definitions:
  InfoResponse:
    type: "object"
    properties:
      typeSubject:
        type: string
      title:
        type: string
      introduction:
        type: string
      publicationDate:
        type: integer
        description: "date in miles seconds"
        example: "1516239022"
      subject:
        type: string
      editorials:
        type: string
      link:
        type: string

  MetadataResponse:
    type: "object"
    properties:
      nameCity:
        type: string
      data:
        type: array
        items:
          $ref: '#/definitions/DataResponse'

  DataResponse:
    type: "object"
    properties:
      description:
        type: string
      unit:
        type: string
      count:
        type: string


  CoordinatesResponse:
    type: "object"
    properties:
      scale:
        type: array
        items:
          type: string

      translate:
        type: array
        items:
          type: string


  ErrorResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"